// src/types/profile.ts
export interface ProfileData {
  uid: string;
  name: string;
  email: string;
  photoURL?: string;
  role: string;
  gender?: string;
  year?: string;
  branch?: string;
  collegeName?: string;
  bio?: string;
  phone?: string;
  team?: string;
  track?: string;
  status?: string;
  github?: string;
  linkedin?: string;
  createdAt: string;
}